refactored create alert so it can be used anywhere
